WiP. Test cases in progress for write_common_metadata and for append_as_row_groups.
